# PERF: Improve performance in to_csv with date format for index