@jreback 2021-12-16
thanks @phofl wouldn't object to addtitional benchmarks as specified in the OP